@@ -1,21 +1,13 @@
-apply plugin: 'com.android.application'
-
+plugins { id 'com.android.application' }
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
-
+    namespace 'com.noti.root'
+    compileSdk 34
     defaultConfig {
         applicationId "com.noti.root"
-        minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 2
-        versionName "1.1"
+        minSdk 24
+        targetSdk 34
+        versionCode 1
+        versionName "1.0"
     }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
+    buildTypes { release { minifyEnabled true; proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro' } }
 }
